#23 code refactoring
